Move all shaper classes into parent file and make file scope
